% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_mediate.R
\name{jm_mediate}
\alias{jm_mediate}
\title{Description of jm_mediate()}
\usage{
jm_mediate(
  jointfit,
  ds_surv,
  ds_long,
  time_eval,
  trt_name,
  time_var,
  L_inner = 50,
  n_mcmc = NULL,
  seed = 15
)
}
\arguments{
\item{jointfit}{A \code{\link[JMbayes2]{jm}} joint model object.}

\item{ds_surv}{Survival data frame that was used to fit the relative risk
submodel.}

\item{ds_long}{Long format data frame that was used to fit the
longitudinal submodel.}

\item{time_eval}{Time points at which to compute effects.}

\item{trt_name}{Name of the treatment variable.}

\item{time_var}{Time variable in the longitudinal model.}

\item{L_inner}{Number of draws for the random effects integration.}

\item{n_mcmc}{How many of the mcmc draws are used.}

\item{seed}{Used for random draws.}
}
\value{
A list containing:
summary: long format data frame of total, direct, and indirect effects over
time;
pop_surv: matrix of survival estimates for
each posterior draw under
s11 = P(do(A^D=1, A^M=1)),
s10 = P(do(A^D=0, A^M=1)),
s01 = P(do(A^D=1, A^M=0)),
s00 = P(do(A^D=0, A^M=0));
pop_cif = 1-pop_surv;
draws: list of effects for each posterior draw;
settings: list of settings used for computation.
}
\description{
Description of jm_mediate()
}
\examples{
# First fit a joint model:
library(JMbayes2)
data("prothro")
data("prothros")

# longitudinal submodel
lmefit <- lme(pro~ ns(time, 3):treat + ns(time,3),
             random = ~time | id,
             data = prothro,
             control = lmeControl(opt = "optim"))
# survival submodel
coxfit <- coxph(Surv(Time, death)~ treat,
            data = prothros, x = TRUE, na.action = na.omit)

# jointmodel
jointfit <- JMbayes2::jm(coxfit, lmefit, time_var = "time", n_chains =2L,
                          n_iter = 40000L, n_burnin = 5000L, n_thin = 3L,
                          data_Surv = prothros)

# Mediation analysis:
med_prothro <- jm_mediate(
                  jointfit = jointfit,
                  ds_surv = prothros,
                  ds_long = prothro,
                  time_eval = 0:12,
                  trt_name = "treat",
                  time_var = "time",
                  n_mcmc = 500
                  )
}
\author{
Lina Kramer \email{linakramer015@gmail.com}
}
